Atlantic Mammals

ATLANTIC MAMMALS: a data set of assemblages of medium- and large-sized mammals of the Atlantic Forest of South America

Usage

atlantic_mammals

Format

A tibble with 4,680 rows and 40 variables:

id

a number indicating the observation

Source

Souza et al. 2019. ATLANTIC MAMMALS: a data set of assemblages of medium- and large-sized mammals of the Atlantic Forest of South America. Ecology. doi:10.1002/ecy.2785