Atlantic Non-volant mammals

Non-volant mammals from the Upper Paraná River Basin: a data set from a critical region for conservation in Brazil

Usage

atlantic_nonvolant_mammals

Format

A tibble with 15,793 rows and 65 variables:

id

a number indicating the observation

Source

Gonçalves et al. 2017. Non-volant mammals from the Upper Paraná River Basin: a data set from a critical region for conservation in Brazil. Ecology, 99(2), 499-499. doi:10.1002/ecy.2107